requirements.txt
pyinstaller
flask
pycryptodome
simple-crypt --no-deps
PySocks
requests
